Move function applyLayerStyleToBillBoardEntity to layer.geojson.controller.ts
